+++ lib/field_struct.rb
@@ -21,6 +21,7 @@ require_relative 'field_struct/types/symbol'
 require_relative 'field_struct/types/uuid'
 require_relative 'field_struct/types/url'
 require_relative 'field_struct/types/email'
+require_relative 'field_struct/types/union'
 require_relative 'field_struct/registry'
 require_relative 'field_struct/field'
 require_relative 'field_struct/metadata'
@@ -51,6 +52,7 @@ module FieldStruct
       r.register :uuid, Types::UUID
       r.register :url, Types::URL
       r.register :email, Types::Email
+      r.register :union, Types::Union
     end
   end
 end